Code/Backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-openai
openai>=1.109.1
httpx>=0.26.0
